requirements.txt
asgiref==3.3.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
boto3==1.17.87
botocore==1.20.87
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==3.4.7
defusedxml==0.6.0
Django==3.2.4
django-allauth==0.44.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-extra-fields==3.0.2
django-filter==2.4.0
django-rest-auth==0.9.5
django-storages==1.11.1
djangorestframework==3.11.0
djangorestframework-jwt
djangorestframework-simplejwt
docutils==0.15.2
drf-friendly-errors==0.14
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
oauthlib==3.1.0
packaging==20.4
Pillow==8.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycparser==2.20
pycryptodome==3.9.9
Pygments==2.6.1
PyJWT
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-mock==3.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
rave-python==1.2.7
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.4.2
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
swagger-spec-validator==2.7.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.12.1
django-phonenumber-field==5.1.0
phonenumbers==8.12.22
django-redis==5.0.0
# django-heroku==0.3.1
dj-database-url==0.5.0
fpdf2==2.4.3